+++ packages/terra-functional-testing/src/commands/expect/index.js
@@ -1,5 +1,7 @@
 const toBeAccessible = require('./toBeAccessible');
+const toMatchReference = require('./toMatchReference');
 
 module.exports = {
   toBeAccessible,
+  toMatchReference,
 };