+++ packages/terra-functional-testing/src/config/wdio.conf.js
@@ -1,12 +1,15 @@
 const fs = require('fs');
 const path = require('path');
-const ip = require('ip');
 const getCapabilities = require('./utils/getCapabilities');
+const getIpAddress = require('./utils/getIpAddress');
 
 const SeleniumDockerService = require('../services/wdio-selenium-docker-service');
 const TerraService = require('../services/wdio-terra-service');
 const AssetServerService = require('../services/wdio-asset-server-service');
-const AccessibilityReporter = require('../reporters/wdio-accessibility-reporter');
+const VisualRegressionLauncher = require('../services/wdio-visual-regression-service');
+
+const { AccessibilityReporter } = require('../reporters/accessibility-reporter');
+const { SpecReporter, cleanResults, mergeResults } = require('../reporters/spec-reporter');
 
 const {
   BROWSERS,
@@ -22,8 +25,6 @@ const {
   WDIO_HOSTNAME,
 } = process.env;
 
-// Convert BROWSERS into an array. When assigned to a process.env it is cast as a string.
-const browsers = BROWSERS ? BROWSERS.split(',') : undefined;
 const defaultWebpackPath = path.resolve(process.cwd(), 'webpack.config.js');
 
 exports.config = {
@@ -70,7 +71,7 @@ exports.config = {
   // Sauce Labs platform configurator - a great tool to configure your capabilities:
   // https://docs.saucelabs.com/reference/platforms-configurator
   //
-  capabilities: getCapabilities(browsers, !!SELENIUM_GRID_URL),
+  capabilities: getCapabilities(BROWSERS, !!SELENIUM_GRID_URL),
   //
   // ===================
   // Test Configurations
@@ -108,7 +109,7 @@ exports.config = {
   // with `/`, the base url gets prepended, not including the path portion of your baseUrl.
   // If your `url` parameter starts without a scheme or `/` (like `some/path`), the base url
   // gets prepended directly.
-  baseUrl: `http://${WDIO_EXTERNAL_HOST || ip.address()}:${WDIO_EXTERNAL_PORT || 8080}`,
+  baseUrl: `http://${WDIO_EXTERNAL_HOST || getIpAddress()}:${WDIO_EXTERNAL_PORT || 8080}`,
   //
   // Default timeout for all waitFor* commands.
   waitforTimeout: 10000,
@@ -137,6 +138,10 @@ exports.config = {
       ...WDIO_INTERNAL_PORT && { port: WDIO_INTERNAL_PORT },
       ...fs.existsSync(defaultWebpackPath) && { webpackConfig: defaultWebpackPath },
     }],
+    [VisualRegressionLauncher, {
+      ...LOCALE && { locale: LOCALE },
+      ...THEME && { theme: THEME },
+    }],
     // Do not add the docker service if disabled.
     ...(WDIO_DISABLE_SELENIUM_SERVICE || SELENIUM_GRID_URL ? [] : [[SeleniumDockerService]]),
   ],
@@ -157,12 +162,27 @@ exports.config = {
   // Test reporter for stdout.
   // The only one supported by default is 'dot'
   // see also: https://webdriver.io/docs/dot-reporter.html
-  reporters: ['spec', AccessibilityReporter],
+  reporters: ['spec', AccessibilityReporter, SpecReporter],
   //
   // Options to be passed to Mocha.
   // See the full list at http://mochajs.org/
   mochaOpts: {
     ui: 'bdd',
     timeout: 60000,
+  },
+  /**
+   * Gets executed once before all workers get launched.
+   */
+  onPrepare() {
+    // Clean previous reporter results.
+    cleanResults();
+  },
+  /**
+   * Gets executed after all workers have shut down and the process is about to exit.
+   * An error thrown in the `onComplete` hook will result in the test run failing.
+   */
+  onComplete() {
+    // Merge reporter results.
+    mergeResults({ formFactor: FORM_FACTOR, locale: LOCALE, theme: THEME });
   },
 };