+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/methods/BaseCompare.js
@@ -1,25 +1,23 @@
 /* eslint-disable class-methods-use-this, no-unused-vars */
-import path from 'path';
+const path = require('path');
 
 /* Pattern to find the shortened test ID within braces. */
 const TEST_ID_REGEX = /\[([^)]+)\]/;
 
-export default class BaseCompare {
+class BaseCompare {
   /**
    * @param {Object} options - Service configuration options.
-   * @param {Object} options.baseScreenshotDir - The base screenshot directory path to save screenshot in.
    * @param {Object} options.locale - The locale being tested.
    * @param {Object} options.theme - The theme being tested.
    */
   constructor(options) {
     const {
-      baseScreenshotDir,
       locale,
       theme,
     } = options;
 
     // screenshot naming config
-    this.baseScreenshotDir = baseScreenshotDir || process.cwd();
+    this.baseScreenshotDir = process.cwd();
     this.locale = locale || 'en';
     this.theme = theme || 'terra-default-theme';
   }
@@ -29,7 +27,7 @@ export default class BaseCompare {
    * the screenshot processing appropriate for the subclass' comparison method.
    *
    * @param {Object} context - Information provided to process the screenshot.
-   * @param {Object} context.browserInfo - Contains the browser's name, version, userAgent.
+   * @param {Object} context.desiredCapabilities - Contains the browser name and capabilities.
    * @param {Object} context.suite - The test suite that is running.
    * @param {Object} context.test - The test that is running.
    * @param {Object} context.meta - Contains the currentFormFactor as meta data to use.
@@ -69,25 +67,25 @@ export default class BaseCompare {
    */
   getScreenshotName(context) {
     const { name } = (context.options || {});
-    const parentName = this.createTestName(context.test.parent);
-    const testName = this.createTestName(name || context.test.title);
+    const testName = this.createTestName(name);
 
-    return `${parentName}[${testName}].png`;
+    return `${testName}.png`;
   }
 
   /**
    * Determines the directories to save the screenshot to.
    *
    * @param {Object} context - Compare context provided by VisualRegressionLauncher.
-   * @param {Object} context.browserInfo - Contains the browser's name, version, userAgent.
+   * @param {Object} context.desiredCapabilities - Contains the browser name and capabilities.
    * @param {Object} context.meta - Contains the currentFormFactor as meta data to use.
    * @returns {String} - The screenshot directory path.
    */
   getScreenshotDir(context) {
-    const { browserInfo, meta } = context;
+    const { desiredCapabilities, meta } = context;
+    const { browserName } = desiredCapabilities;
 
     const formFactor = meta.currentFormFactor;
-    const testForm = `${browserInfo.name}_${formFactor}`;
+    const testForm = `${browserName}_${formFactor}`;
     const testSpec = path.parse(context.test.file).name;
 
     return path.join(this.theme, this.locale, testForm, testSpec);
@@ -121,20 +119,21 @@ export default class BaseCompare {
    *
    * @param {Boolean} referenceExists - Whether or not the screenshot was just created.
    * @param {Number} misMatchPercentage - The percent mismatched of the latest screenshot compared to the reference screenshot.
-   * @param {Boolean} isWithinMisMatchTolerance - Whether or not the latest screenshot is a close enough match the reference screenshot.
+   * @param {Boolean} isWithinMismatchTolerance - Whether or not the latest screenshot is a close enough match the reference screenshot.
    * @param {Boolean} isSameDimensions - Whether or not the latest screenshot was the same dimensions as the reference screenshot.
    * @returns {Object} - The relevant comparison results to report.
    */
-  createResultReport(referenceExists, misMatchPercentage, isWithinMisMatchTolerance, isSameDimensions) {
+  createResultReport(referenceExists, misMatchPercentage, isWithinMismatchTolerance, isSameDimensions) {
     if (!referenceExists) {
       return { isNewScreenshot: true };
     }
 
     return {
       misMatchPercentage,
-      isWithinMisMatchTolerance,
+      isWithinMismatchTolerance,
       isSameDimensions,
-      isExactSameImage: misMatchPercentage === 0 && isSameDimensions,
     };
   }
 }
+
+module.exports = BaseCompare;