+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/scripts/getUserAgent.js
@@ -1,8 +1,0 @@
-/**
- * Script to retrieve the userAgent sent by the browser.
- *
- * @return {string} userAgent
- */
-export default function getUserAgent() {
-  return window.navigator.userAgent;
-}