+++ packages/terra-functional-testing/tests/jest/config/utils/getCapabilities.test.js
@@ -42,4 +42,27 @@ describe('getCapabilities', () => {
     expect(capabilities.length).toEqual(1);
     expect(capabilities[0].browserName).toEqual('internet explorer');
   });
+
+  it('should include only chrome if provided as a string', () => {
+    const capabilities = getCapabilities('chrome');
+
+    expect(capabilities.length).toEqual(1);
+    expect(capabilities[0].browserName).toEqual('chrome');
+  });
+
+  it('should include chrome and firefox when provided as a string', () => {
+    const capabilities = getCapabilities('[chrome, firefox]');
+
+    expect(capabilities.length).toEqual(2);
+    expect(capabilities.findIndex(({ browserName }) => browserName === 'chrome')).toBeGreaterThanOrEqual(0);
+    expect(capabilities.findIndex(({ browserName }) => browserName === 'firefox')).toBeGreaterThanOrEqual(0);
+  });
+
+  it('should include chrome and firefox when provided as a comma delimited  string', () => {
+    const capabilities = getCapabilities('chrome,firefox');
+
+    expect(capabilities.length).toEqual(2);
+    expect(capabilities.findIndex(({ browserName }) => browserName === 'chrome')).toBeGreaterThanOrEqual(0);
+    expect(capabilities.findIndex(({ browserName }) => browserName === 'firefox')).toBeGreaterThanOrEqual(0);
+  });
 });