+++ packages/terra-functional-testing/tests/jest/reporters/accessibility-reporter/index.test.js
@@ -1,0 +1,7 @@
+const AccessibilityReporter = require('../../../../src/reporters/accessibility-reporter');
+
+describe('Accessibility Reporter Index', () => {
+  it('should export the the accessibility reporter', () => {
+    expect(AccessibilityReporter).toBeDefined();
+  });
+});